Create_vectorstores/indexing_requirements.txt
langchain-google-genai
langchain
chromadb 
python-dotenv
langchain_openai